Move API tokenization endpoint into /tokenizers
Requires extracting some more functions out of server.js.
