+++ src/endpoints/tokenizers.js
@@ -6,7 +6,9 @@ const tiktoken = require('@dqbd/tiktoken');
 const { Tokenizer } = require('@agnai/web-tokenizers');
 const { convertClaudePrompt } = require('../chat-completion');
 const { readSecret, SECRET_KEYS } = require('./secrets');
+const { TEXTGEN_TYPES } = require('../constants');
 const { jsonParser } = require('../express-common');
+const { setAdditionalHeaders } = require('../additional-headers');
 
 /**
  * @type {{[key: string]: import("@dqbd/tiktoken").Tiktoken}} Tokenizers cache
@@ -531,6 +533,93 @@ router.post('/openai/count', jsonParser, async function (req, res) {
         const jsonBody = JSON.stringify(req.body);
         const num_tokens = Math.ceil(jsonBody.length / CHARS_PER_TOKEN);
         res.send({ 'token_count': num_tokens });
+    }
+});
+
+router.post('/remote/encode', jsonParser, async function (request, response) {
+    if (!request.body) {
+        return response.sendStatus(400);
+    }
+    const text = String(request.body.text) || '';
+    const api = String(request.body.main_api);
+    const baseUrl = String(request.body.url);
+    const legacyApi = Boolean(request.body.legacy_api);
+
+    try {
+        if (api == 'textgenerationwebui') {
+            const args = {
+                method: 'POST',
+                headers: { 'Content-Type': 'application/json' },
+            };
+
+            setAdditionalHeaders(request, args, null);
+
+            // Convert to string + remove trailing slash + /v1 suffix
+            let url = String(baseUrl).replace(/\/$/, '').replace(/\/v1$/, '');
+
+            if (legacyApi) {
+                url += '/v1/token-count';
+                args.body = JSON.stringify({ 'prompt': text });
+            } else {
+                switch (request.body.api_type) {
+                    case TEXTGEN_TYPES.TABBY:
+                        url += '/v1/token/encode';
+                        args.body = JSON.stringify({ 'text': text });
+                        break;
+                    case TEXTGEN_TYPES.KOBOLDCPP:
+                        url += '/api/extra/tokencount';
+                        args.body = JSON.stringify({ 'prompt': text });
+                        break;
+                    default:
+                        url += '/v1/internal/encode';
+                        args.body = JSON.stringify({ 'text': text });
+                        break;
+                }
+            }
+
+            const result = await fetch(url, args);
+
+            if (!result.ok) {
+                console.log(`API returned error: ${result.status} ${result.statusText}`);
+                return response.send({ error: true });
+            }
+
+            const data = await result.json();
+            const count = legacyApi ? data?.results[0]?.tokens : (data?.length ?? data?.value);
+            const ids = legacyApi ? [] : (data?.tokens ?? []);
+
+            return response.send({ count, ids });
+        }
+
+        else if (api == 'kobold') {
+            const args = {
+                method: 'POST',
+                body: JSON.stringify({ 'prompt': text }),
+                headers: { 'Content-Type': 'application/json' },
+            };
+
+            let url = String(baseUrl).replace(/\/$/, '');
+            url += '/extra/tokencount';
+
+            const result = await fetch(url, args);
+
+            if (!result.ok) {
+                console.log(`API returned error: ${result.status} ${result.statusText}`);
+                return response.send({ error: true });
+            }
+
+            const data = await result.json();
+            const count = data['value'];
+            return response.send({ count: count, ids: [] });
+        }
+
+        else {
+            console.log('Unknown API', api);
+            return response.send({ error: true });
+        }
+    } catch (error) {
+        console.log(error);
+        return response.send({ error: true });
     }
 });
